reports/requirements.txt
APScheduler==3.10.4
arrow==1.3.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
future==1.0.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pure-sasl==0.6.2
PyHive==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
thrift==0.20.0
thrift-sasl==0.4.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
